feat: Initialize React portfolio with essential components and assets
- Added SVG logo and PNG icons for branding.
- Created manifest.json for PWA support.
- Implemented robots.txt for SEO.
- Developed App structure with main components: Navbar, Hero, About, Experience, Skills, Projects, EducationCertificates, Contact, and Footer.
- Styled components with App.css and index.css.
- Set up testing framework with initial App test.
- Included reportWebVitals for performance tracking.
- Established setupTests for testing utilities.
