--- Footer.js
@@ -1,14 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-  return (
-    <footer className="py-8 border-t border-slate-800 text-center text-gray-400 text-sm">
-      <div className="container mx-auto px-6">
-        <p>&copy; 2025 Vicky Nanda Ferdianto. All rights reserved.</p>
-        <p className="mt-2">Designed with ❤️ and modern technologies</p>
-      </div>
-    </footer>
-  );
-};
-
-export default Footer;
